requirements.txt
openai~=2.0
python-dotenv~=1.0
